add UI interface and clic Item Recyclerview

+++ app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/NeighbourFragment.java
@@ -35,16 +35,14 @@ public class NeighbourFragment extends Fragment {
     private NeighbourApiService mApiService;
     //private List<Neighbour> mNeighbours = new Neighbour();
     private List<Neighbour> mNeighbours;
+    private List<Neighbour> mFavoryListe;
     private RecyclerView mRecyclerView;
     private ListNeighbourActivity listNeighbourActivity = new ListNeighbourActivity();
 
     //TODO Aissata
-    Neighbour ActualNeighbourg;
-    private int mPosition;
-   static int mcondition =0;
-   // private List<Neighbour> mFavoriteNeighbours;
-    //List<Neighbour>  mFavoriteNeighbours = new ArrayList<>();
-    List<FavoriteNeighbours> mFavoryListe = new ArrayList<>();
+
+   int mcondition;
+
 
     //FIN
 
@@ -53,11 +51,11 @@ public class NeighbourFragment extends Fragment {
      *
      * @return @{@link NeighbourFragment}
      */
-    public static NeighbourFragment newInstance(boolean condition) {
+    public static NeighbourFragment newInstance(int condition) {
         NeighbourFragment fragment = new NeighbourFragment();
 //TODO AISSATA
         Bundle bundle = new Bundle();
-        bundle.putBoolean("CONDITION", condition);
+        bundle.putInt("CONDITION", condition);
         fragment.setArguments(bundle);
 
      //  mcondition = condition;
@@ -73,7 +71,7 @@ public class NeighbourFragment extends Fragment {
 
         //TODO Aissata
 
-        mcondition =getArguments().getInt("CONDITION", 3);
+        mcondition =getArguments().getInt("CONDITION", 0);
         //FIN
     }
 
@@ -82,14 +80,14 @@ public class NeighbourFragment extends Fragment {
                              Bundle savedInstanceState) {
         //TODO Aissata
 
-        mcondition =getArguments().getInt("CONDITION", 3);
+        mcondition =getArguments().getInt("CONDITION", 0);
         //FIN
         View view = inflater.inflate(R.layout.fragment_neighbour_list, container, false);
         Context context = view.getContext();
         mRecyclerView = (RecyclerView) view;
         mRecyclerView.setLayoutManager(new LinearLayoutManager(context));
         mRecyclerView.addItemDecoration(new DividerItemDecoration(getContext(), DividerItemDecoration.VERTICAL));
-        configureOnClickRecyclerView();
+
 
         return view;
     }
@@ -99,55 +97,66 @@ public class NeighbourFragment extends Fragment {
      */
     //TODO Aissate faire un affichage conditionnel de la liste du recycler view//
     private void initList() {
-       // mNeighbours = mApiService.getNeighbours();
-       // mFavoriteNeighbours.add(mNeighbours.get(2));
-       // mFavoriteNeighbours = new ArrayList<>();
-       // mFavoriteNeighbours = mApiService.getNeighbours();
 
-        //mFavoriteNeighbours.clear();
 // TODO differencier les instancce des listes
         mNeighbours = mApiService.getNeighbours();
+        mFavoryListe =  new ArrayList<>();
+
 
         if(mcondition== 0) {
 
-            //   List< Neighbour> NeighbourTest = new ArrayList<>();
-            // NeighbourTest.add(mNeighbours.get(6));
-            mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours, null));
-            String A = mNeighbours.get(1).getName();
+            mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours));
 
             Log.e("TAG", "initList:  case recyclerview 0 ");
+
         }
 
             else if(mcondition == 1) {
             // mFavoriteNeighbours.add(mNeighbours.get(2));
-            FavoriteNeighbours test = new FavoriteNeighbours(1, "Aissata", mNeighbours.get(2).getAvatarUrl(), "adress 2",
-                    "07 09", "ABOUTme");
-            mFavoryListe.add(test);
+
+            for ( int i = 0; i < mNeighbours.size(); i++){
+                if (mNeighbours.get(i).isFavory()){
+                    mFavoryListe.add(mNeighbours.get(i));
+                }
+
+            }
 
 
-            mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(null, mFavoryListe));
-            // mNeighbours = mApiService.getNeighbours();
-            //mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours,null));
+            mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mFavoryListe));
+
             Log.e("TAG", "initList:  case recyclerview 1 ");
         }
            else{
-              //  mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours,null));
-              //  String A = mNeighbours.get(1).getName();
+
                 Log.e("TAG", "initList:  No RECYCLERVIEW ");
         }
     }
+public int getCondition(){
+    Log.e("TAG", "getCondition: condition: " + mcondition );
+        return mcondition;
+}
 
+
+    @Override
+    public void setUserVisibleHint(boolean isVisibleToUser) {
+        super.setUserVisibleHint(isVisibleToUser);
+        Log.e("TAG", "setUserVisibleHint: " + isVisibleToUser);
+    }
 
     @Override
     public void onResume() {
         super.onResume();
         initList();
+        Log.e("TAG", "onResume: ONRESUME " );
+        getCondition();
     }
 
     @Override
     public void onStart() {
         super.onStart();
         EventBus.getDefault().register(this);
+        Log.e("TAG", "onStart:  ONSTART" );
+        getCondition();
     }
 
     @Override
@@ -161,37 +170,19 @@ public class NeighbourFragment extends Fragment {
      *
      * @param event
      */
+
     @Subscribe
     public void onDeleteNeighbour(DeleteNeighbourEvent event) {
-        mApiService.deleteNeighbour(event.neighbour);
-        initList();
+        if(mcondition ==0) {
+            mApiService.deleteNeighbour(event.neighbour);
+
+            initList();
+        }
     }
 
     //TODO Aissata
 
-    private void configureOnClickRecyclerView() {
 
-        ItemClickSupport.addTo(mRecyclerView).setOnItemClickListener(new ItemClickSupport.OnItemClickListener() {
-            @Override
-            public void onItemClicked(RecyclerView recyclerView, int position, View v) {
-                // do it
-                ActualNeighbourg = mNeighbours.get(position);
-                String NameClicked = ActualNeighbourg.getName();
-                String AvatarClicked = ActualNeighbourg.getAvatarUrl();
-                mPosition = position;
-
-                Log.w("clicked", "onItemClicked: " + position + " nom: " + ActualNeighbourg.getName());
-
-                Intent neigbourgDetailsIntent = new Intent(getActivity(), NeigbourgDetails.class);
-                neigbourgDetailsIntent.putExtra("POSITION", position);
-                neigbourgDetailsIntent.putExtra("NAME_CLICKED", NameClicked);
-                neigbourgDetailsIntent.putExtra("AVATAR_CLICKED", AvatarClicked);
-                startActivity(neigbourgDetailsIntent);
-            }
-        });
-
-
-    }
 
 }
